Add files via upload
Ok, this will need corporation from everyone since we should test three extremity I have each of the procedure 3 times in which one will be the normal the other extreme and last erroneous.  for the procedure that will create a file for class is there three times in the beginning f the test data this is so that the three extremities can be tested.